requirements.txt
asgiref==3.5.2
Django==4.0.5
mysql-connector==2.2.9
sqlparse==0.4.2
tzdata==2022.1
zipfile36==0.1.3
